refactor: GamePrinter 클래스를 별개의 파일로 분리
